Move db and protobuf related code to separate crates
Move `crates/p2p/src/db` to `crates/db`, `crates/p2p/src/wire` to
`crates/wire` and then `crates/wire/proto/` out of `wire` to top level
`proto`. Also some minor refactor stuff and fixes to make CI happy
were applied.
