+++ Cargo.toml
@@ -1,8 +1,42 @@
 [workspace]
-members = ["crates/p2p"]
+members = ["crates/db", "crates/p2p", "crates/wire"]
 
 default-members = ["crates/p2p"]
 
 resolver = "2"
 
 [workspace.dependencies]
+async-trait = "0.1.83"
+bitcoin = { version = "0.32.5", features = ["serde"] }
+futures = "0.3.31"
+libp2p = { version = "0.54.1", features = [
+  "noise",
+  "gossipsub",
+  "tcp",
+  "request-response",
+  "tokio",
+  "secp256k1",
+  "macros",
+  "ping",
+  "yamux",
+  "identify",
+] }
+libp2p-identity = { version = "0.2.10", default-features = false, features = [
+  "secp256k1",
+  "peerid",
+  "rand",
+] }
+musig2 = { version = "0.1.0", features = ["serde"] }
+prost = "0.13.4"
+serde = { version = "1.0.217", features = ["derive"] }
+snafu = { version = "0.8.5", default-features = false, features = [
+  "rust_1_81",
+  "backtrace",
+  "std",
+] }
+tokio = { version = "1.42.0", default-features = false, features = [
+  "macros",
+  "time",
+] }
+tokio-util = { version = "0.7.13", default-features = false }
+tracing = { version = "0.1.41", default-features = false }
