+++ crates/db/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "strata-p2p-db"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+async-trait.workspace = true
+bitcoin.workspace = true
+libp2p-identity.workspace = true
+musig2.workspace = true
+prost.workspace = true
+serde.workspace = true
+snafu.workspace = true
+
+# make this implementation optional
+sled = "0.34.7"
+# temporaty solution for DB serialization. `ciborium` crate is not working, unfortunatly.
+serde_json = "1.0.135"
+
+[dev-dependencies]
+tokio = { version = "1.42.0", features = ["rt", "macros"] }
+rand = "0.8.5"
+secp256k1 = { version = "0.29.0", features = ["rand"] }
