+++ crates/p2p/Cargo.toml
@@ -4,47 +4,36 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-async-trait = "0.1.83"
-bitcoin = { version = "0.32.5", features = ["serde"] }
-futures = "0.3.31"
 libp2p = { version = "0.54.1", features = [
-    "noise",
-    "gossipsub",
-    "tcp",
-    "request-response",
-    "tokio",
-    "secp256k1",
-    "macros",
-    "ping",
-    "yamux",
-    "identify",
+  "noise",
+  "gossipsub",
+  "tcp",
+  "request-response",
+  "tokio",
+  "secp256k1",
+  "macros",
+  "ping",
+  "yamux",
+  "identify",
 ] }
-musig2 = { version = "0.1.0", features = ["serde"] }
-prost = "0.13.4"
-serde = { version = "1.0.217", features = ["derive"] }
-snafu = { version = "0.8.5", default-features = false, features = [
-    "rust_1_81",
-    "backtrace",
-    "std",
-] }
-tokio = { version = "1.42.0", default-features = false, features = [
-    "macros",
-    "time",
-] }
-tokio-util = { version = "0.7.13", default-features = false }
-tracing = { version = "0.1.41", default-features = false }
 
-# make this implementation optional
-sled = "0.34.7"
-# temporaty solution for DB serialization. `ciborium` crate is not working, unfortunatly.
-serde_json = "1.0.135"
+async-trait.workspace = true
+bitcoin = { workspace = true, features = ["serde"] }
+futures.workspace = true
+musig2 = { workspace = true, features = ["serde"] }
+prost.workspace = true
+serde = { workspace = true, features = ["derive"] }
+snafu = { workspace = true, features = ["rust_1_81", "backtrace", "std"] }
+strata-p2p-db.path = "../db"
+strata-p2p-wire.path = "../wire"
+tokio = { workspace = true, features = ["macros", "time"] }
+tokio-util.workspace = true
+tracing.workspace = true
 
 [dev-dependencies]
+sled = "0.34.7"
 tokio = { version = "1.42.0", features = ["rt", "macros", "rt-multi-thread"] }
 rand = "0.8.5"
 secp256k1 = { version = "0.29.0", features = ["rand"] }
 tokio-util = { version = "0.7.13", default-features = false, features = ["rt"] }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
-
-[build-dependencies]
-prost-build = "0.13.4"
