+++ crates/p2p/src/lib.rs
@@ -1,7 +1,4 @@
 pub mod commands;
-pub(crate) mod db;
 pub mod events;
-pub(crate) mod states;
 pub mod swarm;
 pub(crate) mod timeouts;
-pub mod wire;
