--- crates/p2p/src/wire/mod.rs
@@ -1,3 +1,0 @@
-pub mod p2p;
-
-pub use p2p::v1::typed::*;
