--- crates/p2p/src/wire/p2p/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod v1;
