+++ crates/wire/src/p2p/mod.rs
@@ -1,0 +1,1 @@
+pub mod v1;
