+++ crates/wire/src/p2p/v1.rs
@@ -1,0 +1,6 @@
+pub mod proto {
+    include!(concat!(env!("OUT_DIR"), "/strata.bitvm2.p2p.v1.rs"));
+}
+
+pub(crate) mod typed;
+pub use typed::*;
