+++ proto/strata/bitvm2/p2p/v1/getmessage.proto
@@ -27,5 +27,5 @@ message GetMessageRequestBatch {
 }
 
 message GetMessageResponse {
-    repeated GossipsubMsg msg = 1;
+  repeated GossipsubMsg msg = 1;
 }
